requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - Synchronous only
psycopg[binary,pool]==3.1.14
sqlalchemy==2.0.23
alembic==1.12.1

# Redis
redis==5.0.1

# External services
requests==2.31.0
google-generativeai==0.3.0

# API features
slowapi==0.1.9
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0 